Add files via upload
I hope this is the right way to add this file into the notebooks for this repo.  Please let me know if this is correct.  This is a all the EDA I've done thus far - people can edit, organize, remove, etc. things as they want, or copy it and edit on their own.